+++ frontend/src/App.js
@@ -15,6 +15,7 @@ import NotificationContainer from './components/AppNotification';
 import OfflineSyncProgress from './components/OfflineSyncProgress';
 import ConflictResolver from './components/ConflictResolver';
 import PendingUploadsIndicator from './components/PendingUploadsIndicator';
+import PendingSyncChip from './components/PendingSyncChip';
 import { AmberAlertProvider } from './components/AmberAlert';
 import { initErrorReporter, reportError } from './utils/errorReporter';
 import { checkForUpdates } from './utils/versionCheck';
@@ -513,6 +514,7 @@ function App() {
         <SectionLockProvider>
         <BrowserRouter>
           <NetworkStatusBanner />
+          <PendingSyncChip />
           <NotificationContainer />
           <OfflineSyncProgress />
           <PendingUploadsIndicator />